backend/pyproject.toml
[tool.poetry]
name = "CounterfactualCovid19"
version = "0.1.0"
description = ""
authors = [
    "Kellyn Arnold <k.f.arnold@leeds.ac.uk>",
    "Camila Rangel Smith <crangelsmith@turing.ac.uk>",
    "James Robinson <jrobinson@turing.ac.uk>",
]

[tool.poetry.dependencies]
python = "^3.7.9"
Django = "^3.1.5"
django-cors-headers = "^3.6.0"
django-extensions = "^3.1.0"
djangorestframework = "^3.12.2"
djangorestframework-gis = "^0.16"
pandas = "^1.1.5"
psycopg2 = "^2.8.6"
pycountry = "^20.7.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.7.2"
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
